+++ Gemeinsam/BusKomm.h
@@ -51,7 +51,7 @@ extern volatile bool BusEmpfMarkwechsel; //!< \brief Flag für Empfangspegel-Wech
 extern volatile uint16_t TwiIsrCount;
 	//!< Zählt die Anzahl der Aufrufe der TWI-Interrupt-Routine. Nur für Debugging-Zwecke.
 
-extern volatile uint8_t TwiWatchdogCount;
+extern volatile uint16_t TwiWatchdogCount;
 	//!< Wird bei jedem Aufruf der TWI-Interrupt-Routine auf 0 gesetzt. Zur Prüfung der regelmäßigen Kommunikation.
